src/python-server/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
umap-learn>=0.5.3
matplotlib>=3.5.0
seaborn>=0.11.0 